Move MonthUtil to the model as it is used there.
