+++ src/main/java/net/sf/jabref/bibtex/comparator/FieldComparator.java
@@ -18,7 +18,7 @@ package net.sf.jabref.bibtex.comparator;
 import net.sf.jabref.gui.BibtexFields;
 import net.sf.jabref.gui.MainTableFormat;
 import net.sf.jabref.model.entry.AuthorList;
-import net.sf.jabref.logic.util.date.MonthUtil;
+import net.sf.jabref.model.entry.MonthUtil;
 import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.util.Util;
 import net.sf.jabref.logic.util.date.YearUtil;
