+++ src/main/java/net/sf/jabref/gui/actions/CleanUpAction.java
@@ -47,7 +47,7 @@ import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.logic.util.DOI;
 import net.sf.jabref.logic.util.io.FileUtil;
-import net.sf.jabref.logic.util.date.MonthUtil;
+import net.sf.jabref.model.entry.MonthUtil;
 import net.sf.jabref.model.entry.EntryConverter;
 import net.sf.jabref.util.Util;
