+++ src/main/java/net/sf/jabref/logic/xmp/XMPUtil.java
@@ -27,7 +27,7 @@ import net.sf.jabref.importer.ParserResult;
 
 import net.sf.jabref.model.entry.AuthorList;
 import net.sf.jabref.bibtex.BibtexEntryWriter;
-import net.sf.jabref.logic.util.date.MonthUtil;
+import net.sf.jabref.model.entry.MonthUtil;
 import net.sf.jabref.model.database.BibtexDatabase;
 import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.model.entry.BibtexEntryType;
