+++ src/main/java/net/sf/jabref/model/database/BibtexDatabase.java
@@ -29,9 +29,9 @@ Modified for use in JabRef
  */
 package net.sf.jabref.model.database;
 
+import net.sf.jabref.model.entry.MonthUtil;
 import net.sf.jabref.bibtex.EntrySorter;
 import net.sf.jabref.logic.l10n.Localization;
-import net.sf.jabref.logic.util.date.MonthUtil;
 
 import java.beans.PropertyVetoException;
 import java.beans.VetoableChangeListener;
